Move _c structs to cheriabi-specific files.
Remove unused mmsghdr_c entierly.

The relocation of these definitions improves fate-sharing when we
eventually remove CheriABI as a compat layer.
